Rename folder + small tweaks
