+++ apps/mobile/src/App.tsx
@@ -16,7 +16,7 @@ import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { useDeviceContext } from 'twrnc';
 
-import { GlobalModals } from './containers/modals/GlobalModals';
+import { GlobalModals } from './containers/modal/GlobalModals';
 import { reactNativeLink } from './lib/rspcReactNativeTransport';
 import tw from './lib/tailwind';
 import RootNavigator from './navigation';
