+++ apps/mobile/src/navigation/LinkingConfiguration.ts
@@ -1,14 +1,13 @@
-/**
- * Learn more about deep linking with React Navigation
- * https://reactnavigation.org/docs/deep-linking
- * https://reactnavigation.org/docs/configuring-links
- */
 import { LinkingOptions } from '@react-navigation/native';
 import * as Linking from 'expo-linking';
 
 import { RootStackParamList } from '.';
 
-// TODO: Deep linking for React Navigation. It will allow us to do spacedrive://tags/{id} etc.
+/**
+ * TODO: Deep linking for React Navigation. It will allow us to do spacedrive://tags/{id} etc.
+ * https://reactnavigation.org/docs/deep-linking
+ * https://reactnavigation.org/docs/configuring-links
+ */
 const linking: LinkingOptions<RootStackParamList> = {
 	prefixes: [Linking.createURL('/')],
 	config: {
